CVS updated APR

+++ apr/build/apr_common.m4
@@ -102,8 +102,13 @@ changequote([, ])dnl
     ac_sub_cache_file="$ac_popdir/$cache_file" ;;
   esac
 
-  # The eval makes quoting arguments work.
-  if eval $ac_abs_srcdir/configure $ac_configure_args --cache-file=$ac_sub_cache_file --srcdir=$ac_abs_srcdir $2
+  dnl The eval makes quoting arguments work - specifically $2 where the
+  dnl quoting mechanisms used is "" rather than [].
+  dnl
+  dnl We need to execute another shell because some autoconf/shell combinations
+  dnl will choke after doing repeated APR_SUBDIR_CONFIG()s.  (Namely Solaris
+  dnl and autoconf-2.54+)
+  if eval $SHELL $ac_abs_srcdir/configure $ac_configure_args --cache-file=$ac_sub_cache_file --srcdir=$ac_abs_srcdir $2
   then :
     echo "$1 configured properly"
   else
@@ -266,15 +271,18 @@ YES_IS_DEFINED
 ])
 
 dnl
-dnl APR_CHECK_APR_DEFINE( symbol, path_to_apr )
+dnl APR_CHECK_APR_DEFINE( symbol )
 dnl
 AC_DEFUN(APR_CHECK_APR_DEFINE,[
-    AC_EGREP_CPP(YES_IS_DEFINED, [
-#include "$2/include/apr.h"
+apr_old_cppflags=$CPPFLAGS
+CPPFLAGS="$CPPFLAGS $INCLUDES"
+AC_EGREP_CPP(YES_IS_DEFINED, [
+#include <apr.h>
 #if $1
 YES_IS_DEFINED
 #endif
-    ], ac_cv_define_$1=yes, ac_cv_define_$1=no)
+], ac_cv_define_$1=yes, ac_cv_define_$1=no)
+CPPFLAGS=$apr_old_cppflags
 ])
 
 dnl APR_CHECK_FILE(filename); set ac_cv_file_filename to
@@ -454,6 +462,7 @@ AC_DEFUN(APR_CHECK_STRERROR_R_RC,[
 AC_MSG_CHECKING(for type of return code from strerror_r)
 AC_TRY_RUN([
 #include <errno.h>
+#include <string.h>
 #include <stdio.h>
 main()
 {
@@ -707,7 +716,7 @@ dnl that you surround the macro invocation with []s
 AC_DEFUN(APR_HELP_STRING,[ifelse(regexp(AC_ACVERSION, 2\.1), -1, AC_HELP_STRING([$1],[$2]),[  ][$1] substr([                       ],len($1))[$2])])
 
 dnl
-dnl APR_LAYOUT(configlayout, layoutname)
+dnl APR_LAYOUT(configlayout, layoutname [, extravars])
 dnl
 AC_DEFUN(APR_LAYOUT,[
   if test ! -f $srcdir/config.layout; then
@@ -717,19 +726,23 @@ AC_DEFUN(APR_LAYOUT,[
   fi
   pldconf=./config.pld
   changequote({,})
-  sed -e "1,/[ 	]*<[lL]ayout[ 	]*$2[ 	]*>[ 	]*/d" \
+  sed -e "1s/[ 	]*<[lL]ayout[ 	]*$2[ 	]*>[ 	]*//;t" \
+      -e "1,/[ 	]*<[lL]ayout[ 	]*$2[ 	]*>[ 	]*/d" \
       -e '/[ 	]*<\/Layout>[ 	]*/,$d' \
       -e "s/^[ 	]*//g" \
       -e "s/:[ 	]*/=\'/g" \
       -e "s/[ 	]*$/'/g" \
       $1 > $pldconf
   layout_name=$2
+  if test ! -s $pldconf; then
+    echo "** Error: unable to find layout $layout_name"
+    exit 1
+  fi
   . $pldconf
   rm $pldconf
   for var in prefix exec_prefix bindir sbindir libexecdir mandir \
-             sysconfdir datadir  \
-             includedir localstatedir runtimedir logfiledir libdir \
-             installbuilddir libsuffix; do
+             sysconfdir datadir includedir localstatedir runtimedir \
+             logfiledir libdir installbuilddir libsuffix $3; do
     eval "val=\"\$$var\""
     case $val in
       *+)
@@ -759,7 +772,7 @@ AC_DEFUN(APR_LAYOUT,[
 ])dnl
 
 dnl
-dnl APR_ENABLE_LAYOUT(default layout name)
+dnl APR_ENABLE_LAYOUT(default layout name [, extra vars])
 dnl
 AC_DEFUN(APR_ENABLE_LAYOUT,[
 AC_ARG_ENABLE(layout,
@@ -770,7 +783,7 @@ AC_ARG_ENABLE(layout,
 if test -z "$LAYOUT"; then
   LAYOUT="$1"
 fi
-APR_LAYOUT($srcdir/config.layout, $LAYOUT)
+APR_LAYOUT($srcdir/config.layout, $LAYOUT, $2)
 
 AC_MSG_CHECKING(for chosen layout)
 AC_MSG_RESULT($layout_name)
